+++ src/HetrickCV.cpp
@@ -21,6 +21,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelChaos3Op);
 	p->addModel(modelChaoticAttractors);
 	p->addModel(modelClockedNoise);
+	p->addModel(modelClockToPhasor);
 	p->addModel(modelComparator);
 	p->addModel(modelContrast);
 	p->addModel(modelCrackle);