+++ src/HetrickCV.hpp
@@ -42,6 +42,7 @@ extern Model *modelChaos2Op;
 extern Model *modelChaos3Op;
 extern Model *modelChaoticAttractors;
 extern Model *modelClockedNoise;
+extern Model *modelClockToPhasor;
 extern Model *modelComparator;
 extern Model *modelContrast;
 extern Model *modelCrackle;