+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # HetrickCV Changelog
 
-## 2.5.2
+## 2.5.2 (in progress)
+- Add Clock to Phasor.
+- Add Phasor to Random.
 - Add Waveset (in progress).
 - Phasor Swing's default Steps is now "4" for more intuitive behavior.